fix(coding-agent): restore queued abort text
