+++ packages/coding-agent/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 
 - Fixed inherited Anthropic same-model replay to preserve provider-native server tool blocks around signed thinking, avoiding `thinking` / `redacted_thinking` modification errors on follow-up tool-result requests.
+- Fixed interactive abort during streaming or retry to restore queued messages to the editor instead of auto-firing them into a still-settling session.
 
 ### Changed
