+++ packages/coding-agent/test/interactive-mode-abort.test.ts
@@ -35,27 +35,6 @@ function restoreQueuedMessagesToEditor(
 	return restore.call(fakeThis, options);
 }
 
-type AbortAndFireQueuedMessagesThis = {
-	clearAllQueues: () => QueuedMessages;
-	updatePendingMessagesDisplay: () => void;
-	session: {
-		abort: () => Promise<void>;
-		prompt: (text: string) => Promise<void>;
-	};
-	showError: (message: string) => void;
-};
-
-type AbortAndFireQueuedMessages = (this: AbortAndFireQueuedMessagesThis) => Promise<number>;
-
-function abortAndFireQueuedMessages(fakeThis: AbortAndFireQueuedMessagesThis): Promise<number> {
-	const descriptor = Object.getOwnPropertyDescriptor(InteractiveMode.prototype, "abortAndFireQueuedMessages");
-	const fn = descriptor?.value as AbortAndFireQueuedMessages | undefined;
-	if (!fn) {
-		throw new Error("abortAndFireQueuedMessages is missing");
-	}
-	return fn.call(fakeThis);
-}
-
 describe("InteractiveMode.restoreQueuedMessagesToEditor", () => {
 	test("aborts through the session after restoring queued messages", () => {
 		// given
@@ -143,121 +122,5 @@ describe("InteractiveMode.restoreQueuedMessagesToEditor", () => {
 
 		resolveAbort?.();
 		await abortSettled;
-	});
-});
-
-describe("InteractiveMode.abortAndFireQueuedMessages", () => {
-	test("clears queue, aborts, then auto-fires queued messages as a single fresh prompt", async () => {
-		// given
-		const callOrder: string[] = [];
-		const abort = vi.fn<() => Promise<void>>(async () => {
-			callOrder.push("abort");
-		});
-		const prompt = vi.fn<(text: string) => Promise<void>>(async (text: string) => {
-			callOrder.push(`prompt:${text}`);
-		});
-		const updatePendingMessagesDisplay = vi.fn<() => void>();
-		const showError = vi.fn<(message: string) => void>();
-		const fakeThis: AbortAndFireQueuedMessagesThis = {
-			clearAllQueues: () => ({
-				steering: ["msg-a", "msg-b"],
-				followUp: ["msg-c"],
-			}),
-			updatePendingMessagesDisplay,
-			session: { abort, prompt },
-			showError,
-		};
-
-		// when
-		const fired = await abortAndFireQueuedMessages(fakeThis);
-
-		// then
-		expect(fired).toBe(3);
-		expect(callOrder).toEqual(["abort", "prompt:msg-a\n\nmsg-b\n\nmsg-c"]);
-		expect(updatePendingMessagesDisplay).toHaveBeenCalledTimes(1);
-		expect(showError).not.toHaveBeenCalled();
-	});
-
-	test("aborts but does not fire any prompt when no messages are queued", async () => {
-		// given
-		const abort = vi.fn<() => Promise<void>>(async () => {});
-		const prompt = vi.fn<(text: string) => Promise<void>>(async () => {});
-		const updatePendingMessagesDisplay = vi.fn<() => void>();
-		const showError = vi.fn<(message: string) => void>();
-		const fakeThis: AbortAndFireQueuedMessagesThis = {
-			clearAllQueues: () => ({ steering: [], followUp: [] }),
-			updatePendingMessagesDisplay,
-			session: { abort, prompt },
-			showError,
-		};
-
-		// when
-		const fired = await abortAndFireQueuedMessages(fakeThis);
-
-		// then
-		expect(fired).toBe(0);
-		expect(abort).toHaveBeenCalledTimes(1);
-		expect(prompt).not.toHaveBeenCalled();
-		expect(showError).not.toHaveBeenCalled();
-	});
-
-	test("waits for abort to fully settle before firing the queued prompt", async () => {
-		// given
-		let resolveAbort: (() => void) | undefined;
-		const abortSettled = new Promise<void>((resolve) => {
-			resolveAbort = resolve;
-		});
-		const abort = vi.fn<() => Promise<void>>(() => abortSettled);
-		const prompt = vi.fn<(text: string) => Promise<void>>(async () => {});
-		const updatePendingMessagesDisplay = vi.fn<() => void>();
-		const showError = vi.fn<(message: string) => void>();
-		const fakeThis: AbortAndFireQueuedMessagesThis = {
-			clearAllQueues: () => ({ steering: ["queued"], followUp: [] }),
-			updatePendingMessagesDisplay,
-			session: { abort, prompt },
-			showError,
-		};
-
-		// when
-		const promise = abortAndFireQueuedMessages(fakeThis);
-
-		// Yield the microtask queue so the helper reaches its `await session.abort()`.
-		await Promise.resolve();
-		await Promise.resolve();
-
-		// then: prompt has not been called yet because abort hasn't settled
-		expect(prompt).not.toHaveBeenCalled();
-
-		// Resolve abort and let the helper continue
-		resolveAbort?.();
-		const fired = await promise;
-
-		expect(fired).toBe(1);
-		expect(prompt).toHaveBeenCalledWith("queued");
-	});
-
-	test("surfaces fresh prompt errors via showError without throwing", async () => {
-		// given
-		const abort = vi.fn<() => Promise<void>>(async () => {});
-		const prompt = vi.fn<(text: string) => Promise<void>>(async () => {
-			throw new Error("no model selected");
-		});
-		const updatePendingMessagesDisplay = vi.fn<() => void>();
-		const showError = vi.fn<(message: string) => void>();
-		const fakeThis: AbortAndFireQueuedMessagesThis = {
-			clearAllQueues: () => ({ steering: ["msg"], followUp: [] }),
-			updatePendingMessagesDisplay,
-			session: { abort, prompt },
-			showError,
-		};
-
-		// when
-		const fired = await abortAndFireQueuedMessages(fakeThis);
-
-		// then: helper returns the count and reports the error via showError
-		expect(fired).toBe(1);
-		expect(prompt).toHaveBeenCalledWith("msg");
-		expect(showError).toHaveBeenCalledTimes(1);
-		expect(showError.mock.calls[0][0]).toContain("no model selected");
 	});
 });
